leave type desc update

+++ database/migrations/2023_09_19_160402_create_leave_types_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('description')->nullable();
-            $table->integer('period');
+            $table->double('period');
             $table->string('file_date');
             $table->string('code');
             $table->string('attachment')->nullable();
